fixed merge and fixed errors input user
