+++ src/app/components/tweet/tweet.component.html
@@ -1,4 +1,4 @@
 <div class="list">
-  <app-tweet-alone *ngFor="let tweet of tweets" [tweet]="tweet" [user]="user" ></app-tweet-alone>
+  <app-tweet-alone *ngFor="let tweet of tweets" [tweet]="tweet" [me]="me" ></app-tweet-alone>
 </div>
