+++ src/app/components/tweet/tweet.component.ts
@@ -1,11 +1,12 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input, Output } from '@angular/core';
 @Component({
   selector: 'app-tweet',
   templateUrl: './tweet.component.html',
   styleUrls: ['./tweet.component.css']
 })
 export class TweetComponent {
-
   @Input() tweets: Array<any>
-  @Input() user
+  @Input() me
+  ngOnInit() {
+  }
 }
