+++ src/app/components/tweetAlone/tweet-alone/tweet-alone.component.html
@@ -1,5 +1,5 @@
 
-<div class="card p-3" style="min-width: 540px;">
+<div class="card p-3" style="min-width: 540px;" >
   <div class="row no-gutters">
     <div class="col-1">
       <img [src]="user.profilePictureURL" alt="userimages" class="round d-inline card-img" [routerLink]="'/home/' + user.username">
@@ -9,7 +9,7 @@
         <span class="xm-bold" [routerLink]="'/home/' + user.username">{{user.username}}</span>
         <span class="m-medium ml-2" [routerLink]="'/home/' + user.username">@{{user.username}} · </span> <span class="s-medium">{{tweet.creationDate}}</span>
         <p class="card-text">{{ tweet.text }}</p>
-        <img [hidden]="!tweet.urlTweet" [src]="tweet.urlTweet" alt="imgTweet">
+        <img [hidden]="!tweet.urlTweet" [src]="tweet.urlTweet" alt="imgTweet" class="w-100 rounded mx-auto mb-4 cover" height="300px">
             <div class="footer">
               <p class="card-text">RETWEET {{ tweet.numRTs }} </p>
               <p class="card-text">FAVOURITE {{ tweet.numLikes }} </p>
