+++ src/app/pages/home/home.component.html
@@ -1,38 +1,37 @@
+<div class="container">
+  <nav class="row navbar-sm navbar-light fixed-top text-dark m-0 justify-content-center index-up">
+    <div class="col-lg-5">
+      <button id="brand" routerLink="/home" class="btn col-12 col-lg-5 xl-bold green py-2">FRESHH</button>
+      <input class="col-12 col-lg-6 round border p-2 pl-3" type="search" placeholder="Search about topics"
+        maxlength="20" aria-label="Search">
+    </div>
+    <!-- Navtabs -->
+    <div class="tab-nav col-12 col-lg-6 bg-white border">
+      <ul class="nav nav-fill" role="tablist">
+        <li class="nav-item">
+          <a class="nav-link w-100 active show s-bold" id="profile-tab-classic" data-toggle="tab" href="#alltweets"
+            role="tab" aria-controls="alltweets" aria-selected="true">ON LIVE</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link w-100 s-bold" data-toggle="tab" href="#mymeshh" role="tab" aria-controls="mymeshh"
+            aria-selected="false">MY MESHH</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link w-100 s-bold" data-toggle="tab" href="#favs" role="tab" aria-controls="favs"
+            aria-selected="false">FOLLOWING</a>
+        </li>
+      </ul>
+    </div>
+  </nav>
 
-  <div class="container">
-    <nav class="row navbar-sm navbar-light fixed-top text-dark m-0 justify-content-center index-up">
-      <div class="col-lg-5">
-        <button id="brand" routerLink="/home" class="btn col-12 col-lg-5 xl-bold green py-2">FRESHH</button>
-        <input class="col-12 col-lg-6 round border p-2 pl-3" type="search" placeholder="Search about topics"
-          maxlength="20" aria-label="Search">
-      </div>
-      <!-- Navtabs -->
-      <div class="tab-nav col-12 col-lg-6 bg-white border">
-        <ul class="nav nav-fill" role="tablist">
-          <li class="nav-item">
-            <a class="nav-link w-100 active show s-bold" id="profile-tab-classic" data-toggle="tab" href="#alltweets"
-              role="tab" aria-controls="alltweets" aria-selected="true">ON LIVE</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link w-100 s-bold" data-toggle="tab" href="#mymeshh" role="tab" aria-controls="mymeshh"
-              aria-selected="false">MY MESHH</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link w-100 s-bold" data-toggle="tab" href="#favs" role="tab" aria-controls="favs"
-              aria-selected="false">FOLLOWING</a>
-          </li>
-        </ul>
-      </div>
-    </nav>
+</div>
 
-  </div>
+<div class="container-flex justify-self-space bg-light pt-5">
+  <div class="row m-0 justify-content-center">
 
-  <div class="container-flex justify-self-space bg-light pt-5">
-    <div class="row m-0 justify-content-center">
-
-      <div class="col-12 col-lg-5 d-none d-lg-block p-0">
-        <div class="sticky">
-          <app-sidebar [myuser]="user" (propagar)="createTweet($event)"></app-sidebar>
+    <div class="col-12 col-lg-5 d-none d-lg-block p-0">
+      <div class="sticky">
+        <app-sidebar [myuser]="user" (propagar)="createTweet($event)"></app-sidebar>
 
         <div class="col-12 text-center mt-5">
 
@@ -45,19 +44,22 @@
     <div class="col col-lg-6 bg-white max-w p-0">
       <!-- All Tweets -->
       <div class="tab-content col-10 col-lg-12 m-auto p-0" id="myClassicTabContent">
-        <div class="tab-pane fade active show mt-2" id="alltweets" role="tabpanel" aria-labelledby="profile-tab-classic">
-          <app-tweet [tweets]="tweets"></app-tweet>
+        <div class="tab-pane fade active show mt-2" id="alltweets" role="tabpanel"
+          aria-labelledby="profile-tab-classic">
+          <div *ngIf="tweets">
+            <app-tweet [tweets]="tweets" [me]="user"></app-tweet>
+          </div>
         </div>
 
         <!-- My Tweets -->
         <div class="tab-pane fade" id="mymeshh" role="tabpanel" aria-labelledby="follow-tab-classic">
 
-          <app-tweet class="m-5" [tweets]="myTweets" [user]="user"></app-tweet>
+          <app-tweet class="m-5" [tweets]="myTweets" [me]="user"></app-tweet>
         </div>
 
         <!-- Favs -->
         <div class="tab-pane fade" id="favs" role="tabpanel" aria-labelledby="contact-tab-classic">
-          <app-tweet class="m-5" [tweets]="followingtweets" [user]="user"></app-tweet>
+          <app-tweet class="m-5" [tweets]="followingtweets" [me]="user"></app-tweet>
         </div>
       </div>
     </div>
