+++ src/app/pages/otherprofile/otherprofile.component.html
@@ -4,14 +4,14 @@
 </nav>
 
 <div class="container-fluid nav-margin">
-    <div class="tab" id="otheruser"> 
-        <app-otherprofile-user [user]="user" 
+    <div class="tab" id="otheruser">
+        <app-otherprofile-user [user]="user"
         (emitterFollows)="addFollows($event)"
         (emitterNotFollows)="deleteFollows($event)"
         ></app-otherprofile-user>
     </div>
-    
+
     <div class="tab" id="othermeshh">
-        <app-tweet [tweets]="otherTweets" [user]="user"></app-tweet>
+        <app-tweet [tweets]="otherTweets" [me]="user"></app-tweet>
     </div>
 </div>
